updated posts

+++ app/posts/removeRose.php
@@ -2,30 +2,29 @@
 
 require __DIR__.'/../autoload.php';
 
-$postId = filter_var($_GET['rose'], FILTER_SANITIZE_NUMBER_INT);
+if (!userLoggedIn()){
+    redirect('/');
+};
+if (isset($_GET['rose'])) {
 
-$statement = $pdo->prepare('SELECT * FROM posts WHERE id = :id');
-$statement->execute([
-    ':id' => $postId
-]);
+    $postId = filter_var($_GET['rose'], FILTER_SANITIZE_NUMBER_INT);
 
-$postData = $statement->fetch(PDO::FETCH_ASSOC);
+    $postData = getPost($pdo, $postId);
 
-$statement = $pdo->prepare('DELETE FROM roses WHERE post_id = :post_id AND user_id = :user_id');
-$statement->execute([
-    ':user_id' => $_SESSION['user']['id'],
-    ':post_id' => $postId,
-]);
+    $statement = $pdo->prepare('DELETE FROM roses WHERE post_id = :post_id AND user_id = :user_id');
+    $statement->execute([
+        ':user_id' => $_SESSION['user']['id'],
+        ':post_id' => $postId,
+    ]);
 
-$statement = $pdo->prepare('UPDATE posts SET roses = :roses WHERE id = :id');
-$statement->execute([
-    ':roses' => --$postData['roses'],
-    ':id' => $postId
-]);
+    updatePost($pdo, $postId, $postData, 0);
 
-if (!$statement) {
-    die(var_dump($pdo->errorInfo()));
+    if (!$statement) {
+        die(var_dump($pdo->errorInfo()));
+    }
+
+
+    redirect('/post.php?id='.$postId);
 }
 
-
-redirect('/post.php?id='.$postId);
+redirect('/');
